making multisig methods available for other tests
