+++ examples/integration-scripts/multisig-vlei-issuance.test.ts
@@ -1,12 +1,9 @@
 import { strict as assert } from 'assert';
 import signify, {
-    SignifyClient,
     Saider,
-    Serder,
     CredentialSubject,
     CredentialData,
     CreateIdentiferArgs,
-    EventResult,
     randomNonce,
     Salter,
 } from 'signify-ts';
@@ -14,13 +11,25 @@ import { resolveEnvironment } from './utils/resolve-env';
 import {
     resolveOobi,
     waitOperation,
-    waitForNotifications,
+    getOrCreateAID,
+    createTimestamp,
+    getIssuedCredential,
+    getReceivedCredential,
+    waitForCredential,
+    admitSinglesig,
 } from './utils/test-util';
 import { getOrCreateClients, getOrCreateContact } from './utils/test-setup';
 import { HabState } from '../../src/keri/core/state';
-import { retry } from './utils/retry';
-import { step } from './utils/test-step';
-import { addEndRoleMultisig, delegateMultisig } from './utils/multisig-utils';
+import {
+    addEndRoleMultisig,
+    admitMultisig,
+    createAIDMultisig,
+    createRegistryMultisig,
+    delegateMultisig,
+    grantMultisig,
+    issueCredentialMultisig,
+    waitAndMarkNotification,
+} from './utils/multisig-utils';
 
 const { vleiServerUrl, witnessIds } = resolveEnvironment();
 
@@ -69,13 +78,6 @@ const ECR_RULES = Saider.saidify({
         l: 'It is the sole responsibility of Holders as Issuees of an ECR vLEI Credential to present that Credential in a privacy-preserving manner using the mechanisms provided in the Issuance and Presentation Exchange (IPEX) protocol specification and the Authentic Chained Data Container (ACDC) specification. https://github.com/WebOfTrust/IETF-IPEX and https://github.com/trustoverip/tswg-acdc-specification.',
     },
 })[1];
-
-const RETRY_DEFAULTS = {
-    maxSleep: 10000,
-    minSleep: 1000,
-    maxRetries: 10,
-    timeout: 30000,
-};
 
 test('multisig-vlei-issuance', async function run() {
     /**
@@ -1252,397 +1254,4 @@ test('multisig-vlei-issuance', async function run() {
         ecrCredbyECR = await waitForCredential(clientECR, ecrCred.sad.d);
     }
     assert.equal(ecrCred.sad.d, ecrCredbyECR.sad.d);
-}, 3600000);
-
-function createTimestamp() {
-    return new Date().toISOString().replace('Z', '000+00:00');
-}
-
-async function getOrCreateAID(
-    client: SignifyClient,
-    name: string,
-    kargs: CreateIdentiferArgs
-): Promise<HabState> {
-    try {
-        return await client.identifiers().get(name);
-    } catch {
-        const result: EventResult = await client
-            .identifiers()
-            .create(name, kargs);
-
-        await waitOperation(client, await result.op());
-        const aid = await client.identifiers().get(name);
-
-        const op = await client
-            .identifiers()
-            .addEndRole(name, 'agent', client!.agent!.pre);
-        await waitOperation(client, await op.op());
-        console.log(name, 'AID:', aid.prefix);
-        return aid;
-    }
-}
-
-async function createAIDMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    groupName: string,
-    kargs: CreateIdentiferArgs,
-    isInitiator: boolean = false
-) {
-    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/icp');
-
-    const icpResult = await client.identifiers().create(groupName, kargs);
-    const op = await icpResult.op();
-
-    const serder = icpResult.serder;
-    const sigs = icpResult.sigs;
-    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
-    const ims = signify.d(signify.messagize(serder, sigers));
-    const atc = ims.substring(serder.size);
-    const embeds = {
-        icp: [serder, atc],
-    };
-    const smids = kargs.states?.map((state) => state['i']);
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            'multisig',
-            aid,
-            '/multisig/icp',
-            { gid: serder.pre, smids: smids, rmids: smids },
-            embeds,
-            recp
-        );
-
-    return op;
-}
-
-// async function addEndRoleMultisig(
-//     client: SignifyClient,
-//     aid: HabState,
-//     otherMembersAIDs: HabState[],
-//     multisigAID: HabState,
-//     timestamp: string,
-//     isInitiator: boolean = false
-// ) {
-//     if (!isInitiator) await waitAndMarkNotification(client, '/multisig/rpy');
-
-//     const opList: any[] = [];
-//     const members = await client.identifiers().members(multisigAID.name);
-//     const signings = members['signing'];
-
-//     for (const signing of signings) {
-//         const eid = Object.keys(signing.ends.agent)[0];
-//         const endRoleResult = await client
-//             .identifiers()
-//             .addEndRole(multisigAID.name, 'agent', eid, timestamp);
-//         const op = await endRoleResult.op();
-//         opList.push(op);
-
-//         const rpy = endRoleResult.serder;
-//         const sigs = endRoleResult.sigs;
-//         const ghabState1 = multisigAID.state;
-//         const seal = [
-//             'SealEvent',
-//             {
-//                 i: multisigAID.prefix,
-//                 s: ghabState1['ee']['s'],
-//                 d: ghabState1['ee']['d'],
-//             },
-//         ];
-//         const sigers = sigs.map(
-//             (sig: string) => new signify.Siger({ qb64: sig })
-//         );
-//         const roleims = signify.d(
-//             signify.messagize(rpy, sigers, seal, undefined, undefined, false)
-//         );
-//         const atc = roleims.substring(rpy.size);
-//         const roleembeds = {
-//             rpy: [rpy, atc],
-//         };
-//         const recp = otherMembersAIDs.map((aid) => aid.prefix);
-//         await client
-//             .exchanges()
-//             .send(
-//                 aid.name,
-//                 'multisig',
-//                 aid,
-//                 '/multisig/rpy',
-//                 { gid: multisigAID.prefix },
-//                 roleembeds,
-//                 recp
-//             );
-//     }
-
-//     return opList;
-// }
-
-async function createRegistryMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    multisigAID: HabState,
-    registryName: string,
-    nonce: string,
-    isInitiator: boolean = false
-) {
-    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/vcp');
-
-    const vcpResult = await client.registries().create({
-        name: multisigAID.name,
-        registryName: registryName,
-        nonce: nonce,
-    });
-    const op = await vcpResult.op();
-
-    const serder = vcpResult.regser;
-    const anc = vcpResult.serder;
-    const sigs = vcpResult.sigs;
-    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
-    const ims = signify.d(signify.messagize(anc, sigers));
-    const atc = ims.substring(anc.size);
-    const regbeds = {
-        vcp: [serder, ''],
-        anc: [anc, atc],
-    };
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            'registry',
-            aid,
-            '/multisig/vcp',
-            { gid: multisigAID.prefix },
-            regbeds,
-            recp
-        );
-
-    return op;
-}
-
-async function getIssuedCredential(
-    issuerClient: SignifyClient,
-    issuerAID: HabState,
-    recipientAID: HabState,
-    schemaSAID: string
-) {
-    const credentialList = await issuerClient.credentials().list({
-        filter: {
-            '-i': issuerAID.prefix,
-            '-s': schemaSAID,
-            '-a-i': recipientAID.prefix,
-        },
-    });
-    assert(credentialList.length <= 1);
-    return credentialList[0];
-}
-
-async function issueCredentialMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    multisigAIDName: string,
-    kargsIss: CredentialData,
-    isInitiator: boolean = false
-) {
-    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/iss');
-
-    const credResult = await client
-        .credentials()
-        .issue(multisigAIDName, kargsIss);
-    const op = credResult.op;
-
-    const multisigAID = await client.identifiers().get(multisigAIDName);
-    const keeper = client.manager!.get(multisigAID);
-    const sigs = await keeper.sign(signify.b(credResult.anc.raw));
-    const sigers = sigs.map((sig: string) => new signify.Siger({ qb64: sig }));
-    const ims = signify.d(signify.messagize(credResult.anc, sigers));
-    const atc = ims.substring(credResult.anc.size);
-    const embeds = {
-        acdc: [credResult.acdc, ''],
-        iss: [credResult.iss, ''],
-        anc: [credResult.anc, atc],
-    };
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            'multisig',
-            aid,
-            '/multisig/iss',
-            { gid: multisigAID.prefix },
-            embeds,
-            recp
-        );
-
-    return op;
-}
-
-async function grantMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    multisigAID: HabState,
-    recipientAID: HabState,
-    credential: any,
-    timestamp: string,
-    isInitiator: boolean = false
-) {
-    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/exn');
-
-    const [grant, sigs, end] = await client.ipex().grant({
-        senderName: multisigAID.name,
-        acdc: new Serder(credential.sad),
-        anc: new Serder(credential.anc),
-        iss: new Serder(credential.iss),
-        recipient: recipientAID.prefix,
-        datetime: timestamp,
-    });
-
-    await client
-        .ipex()
-        .submitGrant(multisigAID.name, grant, sigs, end, [recipientAID.prefix]);
-
-    const mstate = multisigAID.state;
-    const seal = [
-        'SealEvent',
-        { i: multisigAID.prefix, s: mstate['ee']['s'], d: mstate['ee']['d'] },
-    ];
-    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
-    const gims = signify.d(signify.messagize(grant, sigers, seal));
-    let atc = gims.substring(grant.size);
-    atc += end;
-    const gembeds = {
-        exn: [grant, atc],
-    };
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            'multisig',
-            aid,
-            '/multisig/exn',
-            { gid: multisigAID.prefix },
-            gembeds,
-            recp
-        );
-}
-
-async function admitMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    multisigAID: HabState,
-    recipientAID: HabState,
-    timestamp: string
-    // numGrantMsgs: number
-) {
-    const grantMsgSaid = await waitAndMarkNotification(
-        client,
-        '/exn/ipex/grant'
-    );
-
-    const [admit, sigs, end] = await client
-        .ipex()
-        .admit(multisigAID.name, '', grantMsgSaid, timestamp);
-
-    await client
-        .ipex()
-        .submitAdmit(multisigAID.name, admit, sigs, end, [recipientAID.prefix]);
-
-    const mstate = multisigAID.state;
-    const seal = [
-        'SealEvent',
-        { i: multisigAID.prefix, s: mstate['ee']['s'], d: mstate['ee']['d'] },
-    ];
-    const sigers = sigs.map((sig: string) => new signify.Siger({ qb64: sig }));
-    const ims = signify.d(signify.messagize(admit, sigers, seal));
-    let atc = ims.substring(admit.size);
-    atc += end;
-    const gembeds = {
-        exn: [admit, atc],
-    };
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            'multisig',
-            aid,
-            '/multisig/exn',
-            { gid: multisigAID.prefix },
-            gembeds,
-            recp
-        );
-}
-
-async function admitSinglesig(
-    client: SignifyClient,
-    aid: HabState,
-    recipientAid: HabState
-) {
-    const grantMsgSaid = await waitAndMarkNotification(
-        client,
-        '/exn/ipex/grant'
-    );
-
-    const [admit, sigs, aend] = await client
-        .ipex()
-        .admit(aid.name, '', grantMsgSaid);
-
-    await client
-        .ipex()
-        .submitAdmit(aid.name, admit, sigs, aend, [recipientAid.prefix]);
-}
-
-async function waitAndMarkNotification(client: SignifyClient, route: string) {
-    const notes = await waitForNotifications(client, route);
-
-    await Promise.all(
-        notes.map(async (note) => {
-            await client.notifications().mark(note.i);
-        })
-    );
-
-    return notes[notes.length - 1]?.a.d ?? '';
-}
-
-async function getReceivedCredential(
-    client: SignifyClient,
-    credId: string
-): Promise<any> {
-    const credentialList = await client.credentials().list({
-        filter: {
-            '-d': credId,
-        },
-    });
-    return credentialList[0];
-}
-
-async function waitForCredential(
-    client: SignifyClient,
-    credSAID: string,
-    MAX_RETRIES: number = 10
-) {
-    let retryCount = 0;
-    while (retryCount < MAX_RETRIES) {
-        const cred = await getReceivedCredential(client, credSAID);
-        if (cred) return cred;
-
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        console.log(` retry-${retryCount}: No credentials yet...`);
-        retryCount = retryCount + 1;
-    }
-    throw Error('Credential SAID: ' + credSAID + ' has not been received');
-}
+}, 360000);
