+++ examples/integration-scripts/utils/multisig-utils.ts
@@ -1,5 +1,14 @@
-import signify, { Algos, Siger, SignifyClient, d, messagize } from 'signify-ts';
-import { waitForNotifications, waitOperation } from './test-util';
+import signify, {
+    Algos,
+    CreateIdentiferArgs,
+    CredentialData,
+    Serder,
+    Siger,
+    SignifyClient,
+    d,
+    messagize,
+} from 'signify-ts';
+import { getStates, waitForNotifications, waitOperation } from './test-util';
 import { HabState } from '../../../src/keri/core/state';
 import assert from 'assert';
 
@@ -7,6 +16,17 @@ export interface AcceptMultisigInceptArgs {
     groupName: string;
     localMemberName: string;
     msgSaid: string;
+}
+
+export interface StartMultisigInceptArgs {
+    groupName: string;
+    localMemberName: string;
+    participants: string[];
+    isith?: number | string | string[];
+    nsith?: number | string | string[];
+    toad?: number;
+    wits?: string[];
+    delpre?: string;
 }
 
 export async function acceptMultisigIncept(
@@ -122,15 +142,302 @@ export async function addEndRoleMultisig(
 
     return opList;
 }
-export interface StartMultisigInceptArgs {
-    groupName: string;
-    localMemberName: string;
-    participants: string[];
-    isith?: number | string | string[];
-    nsith?: number | string | string[];
-    toad?: number;
-    wits?: string[];
-    delpre?: string;
+
+export async function admitMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    multisigAID: HabState,
+    recipientAID: HabState,
+    timestamp: string
+    // numGrantMsgs: number
+) {
+    const grantMsgSaid = await waitAndMarkNotification(
+        client,
+        '/exn/ipex/grant'
+    );
+
+    const [admit, sigs, end] = await client
+        .ipex()
+        .admit(multisigAID.name, '', grantMsgSaid, timestamp);
+
+    await client
+        .ipex()
+        .submitAdmit(multisigAID.name, admit, sigs, end, [recipientAID.prefix]);
+
+    const mstate = multisigAID.state;
+    const seal = [
+        'SealEvent',
+        { i: multisigAID.prefix, s: mstate['ee']['s'], d: mstate['ee']['d'] },
+    ];
+    const sigers = sigs.map((sig: string) => new signify.Siger({ qb64: sig }));
+    const ims = signify.d(signify.messagize(admit, sigers, seal));
+    let atc = ims.substring(admit.size);
+    atc += end;
+    const gembeds = {
+        exn: [admit, atc],
+    };
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            'multisig',
+            aid,
+            '/multisig/exn',
+            { gid: multisigAID.prefix },
+            gembeds,
+            recp
+        );
+}
+
+export async function createAIDMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    groupName: string,
+    kargs: CreateIdentiferArgs,
+    isInitiator: boolean = false
+) {
+    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/icp');
+
+    const icpResult = await client.identifiers().create(groupName, kargs);
+    const op = await icpResult.op();
+
+    const serder = icpResult.serder;
+    const sigs = icpResult.sigs;
+    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
+    const ims = signify.d(signify.messagize(serder, sigers));
+    const atc = ims.substring(serder.size);
+    const embeds = {
+        icp: [serder, atc],
+    };
+    const smids = kargs.states?.map((state) => state['i']);
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            'multisig',
+            aid,
+            '/multisig/icp',
+            { gid: serder.pre, smids: smids, rmids: smids },
+            embeds,
+            recp
+        );
+
+    return op;
+}
+
+export async function createRegistryMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    multisigAID: HabState,
+    registryName: string,
+    nonce: string,
+    isInitiator: boolean = false
+) {
+    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/vcp');
+
+    const vcpResult = await client.registries().create({
+        name: multisigAID.name,
+        registryName: registryName,
+        nonce: nonce,
+    });
+    const op = await vcpResult.op();
+
+    const serder = vcpResult.regser;
+    const anc = vcpResult.serder;
+    const sigs = vcpResult.sigs;
+    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
+    const ims = signify.d(signify.messagize(anc, sigers));
+    const atc = ims.substring(anc.size);
+    const regbeds = {
+        vcp: [serder, ''],
+        anc: [anc, atc],
+    };
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            'registry',
+            aid,
+            '/multisig/vcp',
+            { gid: multisigAID.prefix },
+            regbeds,
+            recp
+        );
+
+    return op;
+}
+
+export async function delegateMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    multisigAID: HabState,
+    anchor: { i: string; s: string; d: string },
+    isInitiator: boolean = false
+) {
+    if (!isInitiator) {
+        const msgSaid = await waitAndMarkNotification(client, '/multisig/ixn');
+        console.log(
+            `${aid.name}(${aid.prefix}) received exchange message to join the interaction event`
+        );
+        const res = await client.groups().getRequest(msgSaid);
+        const exn = res[0].exn;
+        const ixn = exn.e.ixn;
+        anchor = ixn.a[0];
+    }
+
+    // const {delResult, delOp} = await retry(async () => {
+    const delResult = await client
+        .delegations()
+        .approve(multisigAID.name, anchor);
+    const appOp = await delResult.op();
+    console.log(
+        `Delegator ${aid.name}(${aid.prefix}) approved delegation for ${
+            multisigAID.name
+        } with anchor ${JSON.stringify(anchor)}`
+    );
+
+    assert.equal(
+        JSON.stringify(delResult.serder.ked.a[0]),
+        JSON.stringify(anchor)
+    );
+
+    const serder = delResult.serder;
+    const sigs = delResult.sigs;
+    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
+    const ims = signify.d(signify.messagize(serder, sigers));
+    const atc = ims.substring(serder.size);
+    const xembeds = {
+        ixn: [serder, atc],
+    };
+    const smids = [aid.prefix, ...otherMembersAIDs.map((aid) => aid.prefix)];
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            multisigAID.name,
+            aid,
+            '/multisig/ixn',
+            { gid: serder.pre, smids: smids, rmids: smids },
+            xembeds,
+            recp
+        );
+
+    if (isInitiator) {
+        console.log(
+            `${aid.name}(${aid.prefix}) initiates delegation interaction event, waiting for others to join...`
+        );
+    } else {
+        console.log(`${aid.name}(${aid.prefix}) joins interaction event`);
+    }
+
+    return appOp;
+}
+
+export async function grantMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    multisigAID: HabState,
+    recipientAID: HabState,
+    credential: any,
+    timestamp: string,
+    isInitiator: boolean = false
+) {
+    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/exn');
+
+    const [grant, sigs, end] = await client.ipex().grant({
+        senderName: multisigAID.name,
+        acdc: new Serder(credential.sad),
+        anc: new Serder(credential.anc),
+        iss: new Serder(credential.iss),
+        recipient: recipientAID.prefix,
+        datetime: timestamp,
+    });
+
+    await client
+        .ipex()
+        .submitGrant(multisigAID.name, grant, sigs, end, [recipientAID.prefix]);
+
+    const mstate = multisigAID.state;
+    const seal = [
+        'SealEvent',
+        { i: multisigAID.prefix, s: mstate['ee']['s'], d: mstate['ee']['d'] },
+    ];
+    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
+    const gims = signify.d(signify.messagize(grant, sigers, seal));
+    let atc = gims.substring(grant.size);
+    atc += end;
+    const gembeds = {
+        exn: [grant, atc],
+    };
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            'multisig',
+            aid,
+            '/multisig/exn',
+            { gid: multisigAID.prefix },
+            gembeds,
+            recp
+        );
+}
+
+export async function issueCredentialMultisig(
+    client: SignifyClient,
+    aid: HabState,
+    otherMembersAIDs: HabState[],
+    multisigAIDName: string,
+    kargsIss: CredentialData,
+    isInitiator: boolean = false
+) {
+    if (!isInitiator) await waitAndMarkNotification(client, '/multisig/iss');
+
+    const credResult = await client
+        .credentials()
+        .issue(multisigAIDName, kargsIss);
+    const op = credResult.op;
+
+    const multisigAID = await client.identifiers().get(multisigAIDName);
+    const keeper = client.manager!.get(multisigAID);
+    const sigs = await keeper.sign(signify.b(credResult.anc.raw));
+    const sigers = sigs.map((sig: string) => new signify.Siger({ qb64: sig }));
+    const ims = signify.d(signify.messagize(credResult.anc, sigers));
+    const atc = ims.substring(credResult.anc.size);
+    const embeds = {
+        acdc: [credResult.acdc, ''],
+        iss: [credResult.iss, ''],
+        anc: [credResult.anc, atc],
+    };
+    const recp = otherMembersAIDs.map((aid) => aid.prefix);
+
+    await client
+        .exchanges()
+        .send(
+            aid.name,
+            'multisig',
+            aid,
+            '/multisig/iss',
+            { gid: multisigAID.prefix },
+            embeds,
+            recp
+        );
+
+    return op;
 }
 
 export async function startMultisigIncept(
@@ -182,13 +489,6 @@ export async function startMultisigIncept(
     return op1;
 }
 
-async function getStates(client: SignifyClient, prefixes: string[]) {
-    const participantStates = await Promise.all(
-        prefixes.map((p) => client.keyStates().get(p))
-    );
-    return participantStates.map((s) => s[0]);
-}
-
 export async function waitAndMarkNotification(
     client: SignifyClient,
     route: string
@@ -202,77 +502,4 @@ export async function waitAndMarkNotification(
     );
 
     return notes[notes.length - 1]?.a.d ?? '';
-}
-
-export async function delegateMultisig(
-    client: SignifyClient,
-    aid: HabState,
-    otherMembersAIDs: HabState[],
-    multisigAID: HabState,
-    anchor: { i: string; s: string; d: string },
-    isInitiator: boolean = false
-) {
-    if (!isInitiator) {
-        const msgSaid = await waitAndMarkNotification(client, '/multisig/ixn');
-        console.log(
-            `${aid.name}(${aid.prefix}) received exchange message to join the interaction event`
-        );
-        const res = await client.groups().getRequest(msgSaid);
-        const exn = res[0].exn;
-        const ixn = exn.e.ixn;
-        anchor = ixn.a[0];
-    }
-
-    // const {delResult, delOp} = await retry(async () => {
-    const delResult = await client
-        .delegations()
-        .approve(multisigAID.name, anchor);
-    const appOp = await delResult.op();
-    console.log(
-        `Delegator ${aid.name}(${aid.prefix}) approved delegation for ${
-            multisigAID.name
-        } with anchor ${JSON.stringify(anchor)}`
-    );
-    // return {delResult, delOp};
-    // },RETRY_DEFAULTS);
-    assert.equal(
-        JSON.stringify(delResult.serder.ked.a[0]),
-        JSON.stringify(anchor)
-    );
-    // const ixnResult = await client
-    //     .identifiers()
-    //     .interact(multisigAID.name, anchor);
-    // const op = await ixnResult.op();
-    const serder = delResult.serder;
-    const sigs = delResult.sigs;
-    const sigers = sigs.map((sig) => new signify.Siger({ qb64: sig }));
-    const ims = signify.d(signify.messagize(serder, sigers));
-    const atc = ims.substring(serder.size);
-    const xembeds = {
-        ixn: [serder, atc],
-    };
-    const smids = [aid.prefix, ...otherMembersAIDs.map((aid) => aid.prefix)];
-    const recp = otherMembersAIDs.map((aid) => aid.prefix);
-
-    await client
-        .exchanges()
-        .send(
-            aid.name,
-            multisigAID.name,
-            aid,
-            '/multisig/ixn',
-            { gid: serder.pre, smids: smids, rmids: smids },
-            xembeds,
-            recp
-        );
-
-    if (isInitiator) {
-        console.log(
-            `${aid.name}(${aid.prefix}) initiates delegation interaction event, waiting for others to join...`
-        );
-    } else {
-        console.log(`${aid.name}(${aid.prefix}) joins interaction event`);
-    }
-
-    return appOp;
 }
